added new copy command & moved config / data function to utils
