+++ biomage/experiment/experiment.py
@@ -8,6 +8,7 @@ from botocore.exceptions import ClientError
 from deepdiff import DeepDiff
 from PyInquirer import prompt
 
+from .copy import copy
 from .ls import ls
 from .pull import pull
 
@@ -227,3 +228,4 @@ def compare(experiment_id):
 experiment.add_command(compare)
 experiment.add_command(pull)
 experiment.add_command(ls)
+experiment.add_command(copy)
